+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -285,7 +285,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K766P1NRMY79SSV14VJ9183C"),
+  _id: z.string().default("01K76WWQR0F7P7RY31Q302C0MB"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -329,7 +329,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K766P1NRMY79SSV14VJ9183C"),
+  id: z.string().default("01K76WWQR0F7P7RY31Q302C0MB"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,