+++ packages/orq-rc/RELEASES.md
@@ -4189,3 +4189,13 @@ Based on:
 - [typescript v3.14.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.19 - packages/orq-rc
+
+## 2025-10-10 11:09:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.20] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.20 - packages/orq-rc